completed/requirements.txt
flask
requests
aiohttp
aiofiles
ngrok
openai
reportlab
pydub
pymongo
alembic
streamlit
streamlit-autorefresh
black
python-dotenv
pytest